sentinel/requirements.txt
dask
planetary_computer
xarray
pystac_client
odc.stac